requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlite3
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0